[fix]: [CDS-79080]: adding deployment freeze changes for custom stage  plan creator
